+++ packages/injectable/index.js
@@ -2,10 +2,17 @@ import getInjectionToken from './src/getInjectionToken/getInjectionToken';
 import getInjectable from './src/getInjectable/getInjectable';
 import lifecycleEnum from './src/dependency-injection-container/lifecycleEnum';
 import createContainer from './src/dependency-injection-container/createContainer';
+
 import {
   registerDependencyGraphing,
   plantUmlDependencyGraphInjectable,
+  dependencyGraphCustomizerToken,
 } from './src/dependency-injection-container/extensions/dependency-graphing/dependency-graphing';
+
+import {
+  registerErrorMonitoring,
+  errorMonitorInjectionToken,
+} from './src/dependency-injection-container/extensions/error-monitoring/error-monitoring';
 
 export {
   getInjectionToken,
@@ -14,4 +21,7 @@ export {
   createContainer,
   registerDependencyGraphing,
   plantUmlDependencyGraphInjectable,
+  dependencyGraphCustomizerToken,
+  registerErrorMonitoring,
+  errorMonitorInjectionToken,
 };