+++ packages/injectable/src/dependency-injection-container/createContainer.dependency-graphing.test.js
@@ -1,9 +1,13 @@
 import getDi from '../test-utils/getDiForUnitTesting';
 import getInjectable from '../getInjectable/getInjectable';
+
 import {
+  dependencyGraphCustomizerToken,
   plantUmlDependencyGraphInjectable,
   registerDependencyGraphing,
 } from './extensions/dependency-graphing/dependency-graphing';
+
+import isEqual from 'lodash/fp/isEqual';
 import getInjectionToken from '../getInjectionToken/getInjectionToken';
 import lifecycleEnum from './lifecycleEnum';
 
@@ -12,43 +16,106 @@ describe('createContainer.dependency-graph', () => {
     const parentInjectable = getInjectable({
       id: 'some-parent-injectable',
 
-      instantiate: di => di.inject(childInjectable),
+      instantiate: async di => {
+        // Inject same injectable twice for coverage
+        di.inject(syncChildInjectable);
+        di.inject(syncChildInjectable);
+        await di.inject(asyncChildInjectable);
+        di.inject(keyedInjectable);
+      },
     });
 
-    const childInjectable = getInjectable({
-      id: 'some-child-injectable',
+    const syncChildInjectable = getInjectable({
+      id: 'some-sync-child-injectable',
       instantiate: di => di.injectMany(injectionToken),
       lifecycle: lifecycleEnum.transient,
     });
 
+    const asyncChildInjectable = getInjectable({
+      id: 'some-async-child-injectable',
+      instantiate: async di => {
+        di.inject(customizableAsyncInjectable);
+
+        return 'irrelevant';
+      },
+    });
+
+    const keyedInjectable = getInjectable({
+      id: 'some-keyed-injectable',
+      instantiate: () => 'irrelevant',
+      lifecycle: lifecycleEnum.keyedSingleton({
+        getInstanceKey: () => 'irrelevant',
+      }),
+    });
+
     const injectionToken = getInjectionToken({ id: 'some-injection-token' });
+
     const tokenInjectable = getInjectable({
       id: 'some-token-injectable',
       instantiate: () => 'irrelevant',
       injectionToken,
     });
 
-    const setuppableInjectable = getInjectable({
+    const customizableSyncInjectable = getInjectable({
+      id: 'some-customizable-sync-injectable',
+      instantiate: () => 'some-customizable-instance',
+      injectionToken,
+    });
+
+    const customizableAsyncInjectable = getInjectable({
+      id: 'some-customizable-async-injectable',
+      instantiate: async () => 'some-customizable-instance',
+    });
+
+    const graphCustomizer = getInjectable({
+      id: 'some-dependency-graph-customizer',
+
+      instantiate: () => ({
+        shouldCustomize: isEqual('some-customizable-instance'),
+
+        customizeLink: link => {
+          link.infos.add('some-custom-link-info');
+          link.lineColor = 'orange';
+          link.textColor = 'green';
+        },
+
+        customizeNode: node => {
+          node.infos.add('some-custom-node-info');
+        },
+      }),
+
+      decorable: false,
+
+      injectionToken: dependencyGraphCustomizerToken,
+    });
+
+    const setuppable = getInjectable({
       id: 'some-setuppable',
       instantiate: () => 'irrelevant',
-      setup: di => {
-        di.inject(childInjectable);
-        di.inject(setuppableInjectable);
+
+      setup: async di => {
+        await di.inject(syncChildInjectable);
+        await di.inject(setuppable);
       },
     });
 
     const di = getDi(
       parentInjectable,
-      childInjectable,
+      syncChildInjectable,
+      asyncChildInjectable,
       tokenInjectable,
-      setuppableInjectable,
+      setuppable,
+      keyedInjectable,
+      customizableSyncInjectable,
+      customizableAsyncInjectable,
+      graphCustomizer,
     );
 
     registerDependencyGraphing(di);
 
     await di.runSetups();
 
-    di.inject(parentInjectable);
+    await di.inject(parentInjectable);
 
     const graph = di.inject(plantUmlDependencyGraphInjectable);