+++ packages/injectable/src/dependency-injection-container/createContainer.error-monitoring-for-injected-functions.test.js
@@ -2,7 +2,7 @@ import asyncFn from '@async-fn/jest';
 import getDi from '../test-utils/getDiForUnitTesting';
 import getInjectable from '../getInjectable/getInjectable';
 import lifecycleEnum from './lifecycleEnum';
-import { noop } from 'lodash/fp';
+import noop from 'lodash/fp/noop';
 
 import {
   errorMonitorInjectionToken,
@@ -24,6 +24,7 @@ describe('createContainer.error-monitoring-for-injected-functions', () => {
         id: 'some-error-monitor',
         injectionToken: errorMonitorInjectionToken,
         instantiate: () => errorMonitorMock,
+        decorable: false,
       });
 
       childCallMock = asyncFn();
@@ -129,6 +130,7 @@ describe('createContainer.error-monitoring-for-injected-functions', () => {
         id: 'some-error-monitor',
         injectionToken: errorMonitorInjectionToken,
         instantiate: () => errorMonitorMock,
+        decorable: false,
       });
 
       childCallMock = asyncFn();