+++ packages/injectable/src/dependency-injection-container/createContainer.setuppable.test.js
@@ -3,9 +3,10 @@ import lifecycleEnum from './lifecycleEnum';
 import getInjectable from '../getInjectable/getInjectable';
 import getPromiseStatus from '../test-utils/getPromiseStatus/getPromiseStatus';
 import getDi from '../test-utils/getDiForUnitTesting';
+import getInjectionToken from '../getInjectionToken/getInjectionToken';
 
 describe('createContainer.setuppable', () => {
-  it('given setuppables with a dependency cycle, when setupped, throws most complex cycle in system', () => {
+  it('given setuppables with a dependency cycle when injecting single, when setupped, throws most complex cycle in system', () => {
     const someRootSetuppable = getInjectable({
       id: 'some-root-injectable',
       setup: async di => {
@@ -39,14 +40,49 @@ describe('createContainer.setuppable', () => {
     );
   });
 
-  it('given setup for injectable, when setups are ran, runs the setup with a way to inject', async () => {
+  xit('given setuppables with a dependency cycle when injecting many, when setupped, throws most complex cycle in system', async () => {
+    const someToken = getInjectionToken({ id: 'some-token' });
+    const someSetuppable = getInjectable({
+      id: 'some-injectable',
+      setup: async di => {
+        await di.injectMany(someOtherToken);
+      },
+
+      instantiate: () => 'irrelevant',
+
+      injectionToken: someToken,
+    });
+
+    const someOtherToken = getInjectionToken({ id: 'some-other-token' });
+    const someOtherSetuppable = getInjectable({
+      id: 'some-other-injectable',
+      setup: async di => {
+        await di.injectMany(someToken);
+      },
+
+      instantiate: () => 'irrelevant',
+
+      injectionToken: someOtherToken,
+    });
+
+    const di = getDi(someSetuppable, someOtherSetuppable);
+
+    return expect(di.runSetups()).rejects.toThrow(
+      'Cycle of setuppables encountered: "some-root-injectable" -> "some-parent-injectable" -> "some-child-injectable" -> "some-parent-injectable"',
+    );
+  });
+
+  it('given setup for injectable, when setups are ran, runs the setup with a ways to inject', async () => {
     let instanceFromSetup;
 
     const someSetuppable = getInjectable({
       id: 'some-setuppable',
 
       setup: async di => {
-        instanceFromSetup = await di.inject(someInjectable, 'some-parameter');
+        instanceFromSetup = {
+          single: await di.inject(someInjectable, 'some-parameter'),
+          many: await di.injectMany(someInjectionToken, 'some-parameter'),
+        };
       },
 
       instantiate: () => 'irrelavent',
@@ -55,14 +91,28 @@ describe('createContainer.setuppable', () => {
     const someInjectable = getInjectable({
       id: 'some-injectable',
       lifecycle: lifecycleEnum.transient,
-      instantiate: (di, parameter) => `some-instance: "${parameter}"`,
+      instantiate: (di, parameter) => `some-single-instance: "${parameter}"`,
     });
 
-    const di = getDi(someSetuppable, someInjectable);
+    const someInjectionToken = getInjectionToken({
+      id: 'some-injection-token',
+    });
+
+    const someTokenInjectable = getInjectable({
+      id: 'some-token-injectable',
+      lifecycle: lifecycleEnum.transient,
+      instantiate: (di, parameter) => `some-of-many-instances: "${parameter}"`,
+      injectionToken: someInjectionToken,
+    });
+
+    const di = getDi(someSetuppable, someInjectable, someTokenInjectable);
 
     await di.runSetups();
 
-    expect(instanceFromSetup).toBe('some-instance: "some-parameter"');
+    expect(instanceFromSetup).toEqual({
+      single: 'some-single-instance: "some-parameter"',
+      many: ['some-of-many-instances: "some-parameter"'],
+    });
   });
 
   it('given multiple async setuppables and setups are ran, when setups resolve, setup resolves', async () => {